+++ tests/app/test_dependencies.py
@@ -189,9 +189,6 @@ def test_get_agent(monkeypatch, patch_required_env):
     settings = Settings()
 
     language_model = get_language_model(settings)
-    bluenaas_tool = get_bluenaas_tool(
-        settings=settings, token=token, httpx_client=httpx_client
-    )
     literature_tool = get_literature_tool(
         token=token, settings=settings, httpx_client=httpx_client
     )
@@ -221,7 +218,6 @@ def test_get_agent(monkeypatch, patch_required_env):
 
     agent = get_agent(
         llm=language_model,
-        bluenaas_tool=bluenaas_tool,
         literature_tool=literature_tool,
         br_resolver_tool=br_resolver_tool,
         morpho_tool=morpho_tool,
@@ -250,6 +246,9 @@ async def test_get_chat_agent(monkeypatch, db_connection, patch_required_env):
     )
     literature_tool = get_literature_tool(
         token=token, settings=settings, httpx_client=httpx_client
+    )
+    me_model_tool = get_me_model_tool(
+        settings=settings, token=token, httpx_client=httpx_client
     )
     morpho_tool = get_morpho_tool(
         settings=settings, token=token, httpx_client=httpx_client
@@ -281,6 +280,7 @@ async def test_get_chat_agent(monkeypatch, db_connection, patch_required_env):
         br_resolver_tool=br_resolver_tool,
         morpho_tool=morpho_tool,
         morphology_feature_tool=morphology_feature_tool,
+        me_model_tool=me_model_tool,
         kg_morpho_feature_tool=kg_morpho_feature_tool,
         electrophys_feature_tool=electrophys_feature_tool,
         traces_tool=traces_tool,